+++ app/keepers/keepers.go
@@ -60,6 +60,7 @@ import (
 	ibckeeper "github.com/cosmos/ibc-go/v8/modules/core/keeper"
 
 	appparams "github.com/babylonlabs-io/babylon/app/params"
+	"github.com/babylonlabs-io/babylon/app/signer"
 	bbn "github.com/babylonlabs-io/babylon/types"
 	owasm "github.com/babylonlabs-io/babylon/wasmbinding"
 	btccheckpointkeeper "github.com/babylonlabs-io/babylon/x/btccheckpoint/keeper"
@@ -162,7 +163,7 @@ func (ak *AppKeepers) InitKeepers(
 	homePath string,
 	invCheckPeriod uint,
 	skipUpgradeHeights map[int64]bool,
-	privSigner *PrivSigner,
+	privSigner *signer.PrivSigner,
 	appOpts servertypes.AppOptions,
 	wasmConfig wasmtypes.WasmConfig,
 	wasmOpts []wasmkeeper.Option,
@@ -217,7 +218,7 @@ func (ak *AppKeepers) InitKeepers(
 		maccPerms,
 		authcodec.NewBech32Codec(appparams.Bech32PrefixAccAddr),
 		appparams.Bech32PrefixAccAddr,
-		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
+		appparams.AccGov.String(),
 	)
 
 	bankKeeper := bankkeeper.NewBaseKeeper(
@@ -225,7 +226,7 @@ func (ak *AppKeepers) InitKeepers(
 		runtime.NewKVStoreService(keys[banktypes.StoreKey]),
 		accountKeeper,
 		blockedAddress,
-		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
+		appparams.AccGov.String(),
 		logger,
 	)
 
@@ -234,7 +235,7 @@ func (ak *AppKeepers) InitKeepers(
 		runtime.NewKVStoreService(keys[stakingtypes.StoreKey]),
 		accountKeeper,
 		bankKeeper,
-		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
+		appparams.AccGov.String(),
 		authcodec.NewBech32Codec(appparams.Bech32PrefixValAddr),
 		authcodec.NewBech32Codec(appparams.Bech32PrefixConsAddr),
 	)
@@ -245,7 +246,7 @@ func (ak *AppKeepers) InitKeepers(
 		runtime.NewKVStoreService(keys[epochingtypes.StoreKey]),
 		bankKeeper,
 		stakingKeeper,
-		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
+		appparams.AccGov.String(),
 	)
 
 	checkpointingKeeper := checkpointingkeeper.NewKeeper(
@@ -270,7 +271,7 @@ func (ak *AppKeepers) InitKeepers(
 	ak.ConsensusParamsKeeper = consensusparamkeeper.NewKeeper(
 		appCodec,
 		runtime.NewKVStoreService(keys[consensusparamtypes.StoreKey]),
-		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
+		appparams.AccGov.String(),
 		runtime.EventService{},
 	)
 	bApp.SetParamStore(ak.ConsensusParamsKeeper.ParamsStore)
@@ -300,7 +301,7 @@ func (ak *AppKeepers) InitKeepers(
 	ak.CircuitKeeper = circuitkeeper.NewKeeper(
 		appCodec,
 		runtime.NewKVStoreService(keys[circuittypes.StoreKey]),
-		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
+		appparams.AccGov.String(),
 		ak.AccountKeeper.AddressCodec(),
 	)
 	bApp.SetCircuitBreaker(&ak.CircuitKeeper)
@@ -312,7 +313,7 @@ func (ak *AppKeepers) InitKeepers(
 		ak.AccountKeeper,
 		ak.BankKeeper,
 		authtypes.FeeCollectorName,
-		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
+		appparams.AccGov.String(),
 	)
 
 	ak.DistrKeeper = distrkeeper.NewKeeper(
@@ -322,7 +323,7 @@ func (ak *AppKeepers) InitKeepers(
 		ak.BankKeeper,
 		ak.StakingKeeper,
 		authtypes.FeeCollectorName,
-		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
+		appparams.AccGov.String(),
 	)
 
 	// set up incentive keeper
@@ -332,7 +333,7 @@ func (ak *AppKeepers) InitKeepers(
 		ak.BankKeeper,
 		ak.AccountKeeper,
 		&epochingKeeper,
-		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
+		appparams.AccGov.String(),
 		authtypes.FeeCollectorName,
 	)
 
@@ -341,7 +342,7 @@ func (ak *AppKeepers) InitKeepers(
 		encodingConfig.Amino,
 		runtime.NewKVStoreService(keys[slashingtypes.StoreKey]),
 		ak.StakingKeeper,
-		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
+		appparams.AccGov.String(),
 	)
 
 	ak.CrisisKeeper = crisiskeeper.NewKeeper(
@@ -350,7 +351,7 @@ func (ak *AppKeepers) InitKeepers(
 		invCheckPeriod,
 		ak.BankKeeper,
 		authtypes.FeeCollectorName,
-		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
+		appparams.AccGov.String(),
 		ak.AccountKeeper.AddressCodec(),
 	)
 
@@ -372,7 +373,7 @@ func (ak *AppKeepers) InitKeepers(
 		appCodec,
 		homePath,
 		bApp,
-		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
+		appparams.AccGov.String(),
 	)
 
 	ak.AuthzKeeper = authzkeeper.NewKeeper(
@@ -393,7 +394,7 @@ func (ak *AppKeepers) InitKeepers(
 		// `MsgIBCSoftwareUpgrade` and `MsgRecoverClient`
 		// https://github.com/cosmos/ibc-go/releases/tag/v8.0.0
 		// Gov is the proper authority for those types of messages
-		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
+		appparams.AccGov.String(),
 	)
 
 	// register the proposal types
@@ -420,7 +421,7 @@ func (ak *AppKeepers) InitKeepers(
 		ak.DistrKeeper,
 		bApp.MsgServiceRouter(),
 		govConfig,
-		authtypes.NewModuleAddress(govtypes.ModuleName).String())
+		appparams.AccGov.String())
 
 	ak.GovKeeper = *govKeeper.SetHooks(
 		govtypes.NewMultiGovHooks(
@@ -433,7 +434,7 @@ func (ak *AppKeepers) InitKeepers(
 		runtime.NewKVStoreService(keys[btclightclienttypes.StoreKey]),
 		*btcConfig,
 		&ak.IncentiveKeeper,
-		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
+		appparams.AccGov.String(),
 	)
 
 	btcCheckpointKeeper := btccheckpointkeeper.NewKeeper(
@@ -444,7 +445,7 @@ func (ak *AppKeepers) InitKeepers(
 		&checkpointingKeeper,
 		&ak.IncentiveKeeper,
 		&powLimit,
-		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
+		appparams.AccGov.String(),
 	)
 
 	ak.IBCFeeKeeper = ibcfeekeeper.NewKeeper(
@@ -465,7 +466,7 @@ func (ak *AppKeepers) InitKeepers(
 		ak.AccountKeeper,
 		ak.BankKeeper,
 		scopedTransferKeeper,
-		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
+		appparams.AccGov.String(),
 	)
 
 	ak.MonitorKeeper = monitorkeeper.NewKeeper(
@@ -498,7 +499,7 @@ func (ak *AppKeepers) InitKeepers(
 		&btcCheckpointKeeper,
 		&ak.IncentiveKeeper,
 		btcNetParams,
-		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
+		appparams.AccGov.String(),
 	)
 
 	// set up finality keeper
@@ -508,7 +509,7 @@ func (ak *AppKeepers) InitKeepers(
 		ak.BTCStakingKeeper,
 		ak.IncentiveKeeper,
 		ak.CheckpointingKeeper,
-		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
+		appparams.AccGov.String(),
 	)
 
 	// create evidence keeper with router
@@ -542,7 +543,7 @@ func (ak *AppKeepers) InitKeepers(
 		homePath,
 		wasmConfig,
 		WasmCapabilities(),
-		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
+		appparams.AccGov.String(),
 		wasmOpts...,
 	)
 
@@ -557,7 +558,7 @@ func (ak *AppKeepers) InitKeepers(
 		appCodec,
 		runtime.NewKVStoreService(keys[ibcwasmtypes.StoreKey]),
 		ak.IBCKeeper.ClientKeeper,
-		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
+		appparams.AccGov.String(),
 		ibcWasmConfig,
 		bApp.GRPCQueryRouter(),
 	)